Move models and collections to server dir in api
