+++ api/server/main.ts
@@ -2,7 +2,7 @@ import { Meteor } from 'meteor/meteor';
 import { Accounts } from 'meteor/accounts-base';
 import { initMethods } from './methods';
 import { initPublications } from './publications';
-import { Users } from '../collections/whatsapp';
+import { Users } from './collections/whatsapp';
 
 Meteor.startup(() => {
   initMethods();
