+++ api/server/methods.ts
@@ -3,8 +3,8 @@ import { check, Match } from 'meteor/check';
 import { UploadFS } from 'meteor/jalik:ufs';
 import { Profile } from 'api/models/whatsapp';
 import { Picture } from 'api/models/ufs';
-import { Chats, Messages } from '../collections/whatsapp';
-import { Pictures } from '../collections/ufs';
+import { Chats, Messages } from './collections/whatsapp';
+import { Pictures } from './collections/ufs';
 
 const nonEmptyString = Match.Where((str) => {
   check(str, String);
